+++ common/windows/cpp/include/ConvertUTF.h
