+++ common/windows/cpp/include/crc32.h
