+++ common/windows/cpp/include/keymansentry.h
