+++ common/windows/cpp/include/registry.h
