+++ common/windows/cpp/include/xstring.h
