+++ common/windows/cpp/src/ConvertUTF.c
@@ -1,8 +1,8 @@
 /*
  * Copyright 2001 Unicode, Inc.
- * 
+ *
  * Disclaimer
- * 
+ *
  * This source code is provided as is by Unicode, Inc. No claims are
  * made as to fitness for any particular purpose. No warranties of any
  * kind are expressed or implied. The recipient agrees to determine
@@ -10,9 +10,9 @@
  * purchased on magnetic or optical media from Unicode, Inc., the
  * sole remedy for any claim will be exchange of defective media
  * within 90 days of receipt.
- * 
+ *
  * Limitations on Rights to Redistribute This Code
- * 
+ *
  * Unicode, Inc. hereby grants the right to freely use the information
  * supplied in this file in the creation of products supporting the
  * Unicode Standard, and to make copies of this file in any form
@@ -30,7 +30,7 @@
 
 ------------------------------------------------------------------------ */
 
-#include "ConvertUTF.h"
+#include "../include/ConvertUTF.h"
 #ifdef CVTUTF_DEBUG
 #include <stdio.h>
 #endif
@@ -50,7 +50,7 @@ static const UTF32 halfMask	= 0x3FFUL;
 /* --------------------------------------------------------------------- */
 
 ConversionResult ConvertUTF32toUTF16 (
-		UTF32** sourceStart, const UTF32* sourceEnd, 
+		UTF32** sourceStart, const UTF32* sourceEnd,
 		UTF16** targetStart, const UTF16* targetEnd, const ConversionFlags flags) {
 	ConversionResult result = conversionOK;
 	UTF32* source = *sourceStart;
@@ -93,7 +93,7 @@ ConversionResult ConvertUTF32toUTF16 (
 /* --------------------------------------------------------------------- */
 
 ConversionResult ConvertUTF16toUTF32 (
-		UTF16** sourceStart, UTF16* sourceEnd, 
+		UTF16** sourceStart, UTF16* sourceEnd,
 		UTF32** targetStart, const UTF32* targetEnd, const ConversionFlags flags) {
 	ConversionResult result = conversionOK;
 	UTF16* source = *sourceStart;
@@ -156,7 +156,7 @@ static const char trailingBytesForUTF8[256] = {
  * This table contains as many values as there might be trailing bytes
  * in a UTF-8 sequence.
  */
-static const UTF32 offsetsFromUTF8[6] = { 0x00000000UL, 0x00003080UL, 0x000E2080UL, 
+static const UTF32 offsetsFromUTF8[6] = { 0x00000000UL, 0x00003080UL, 0x000E2080UL,
 					 0x03C82080UL, 0xFA082080UL, 0x82082080UL };
 
 /*
@@ -180,7 +180,7 @@ static const UTF8 firstByteMark[7] = { 0x00, 0x00, 0xC0, 0xE0, 0xF0, 0xF8, 0xFC 
 /* --------------------------------------------------------------------- */
 
 ConversionResult ConvertUTF16toUTF8 (
-		UTF16** sourceStart, const UTF16* sourceEnd, 
+		UTF16** sourceStart, const UTF16* sourceEnd,
 		UTF8** targetStart, const UTF8* targetEnd, const ConversionFlags flags) {
 	ConversionResult result = conversionOK;
 	UTF16* source = *sourceStart;
@@ -189,7 +189,7 @@ ConversionResult ConvertUTF16toUTF8 (
 		UTF32 ch;
 		unsigned short bytesToWrite = 0;
 		const UTF32 byteMask = 0xBF;
-		const UTF32 byteMark = 0x80; 
+		const UTF32 byteMark = 0x80;
 		ch = *source++;
 		/* If we have a surrogate pair, convert to UTF32 first. */
 		if (ch >= UNI_SUR_HIGH_START && ch <= UNI_SUR_HIGH_END && source < sourceEnd) {
@@ -286,7 +286,7 @@ Boolean isLegalUTF8Sequence(UTF8 *source, UTF8 *sourceEnd) {
 /* --------------------------------------------------------------------- */
 
 ConversionResult ConvertUTF8toUTF16 (
-		UTF8** sourceStart, UTF8* sourceEnd, 
+		UTF8** sourceStart, UTF8* sourceEnd,
 		UTF16** targetStart, const UTF16* targetEnd, const ConversionFlags flags) {
 	ConversionResult result = conversionOK;
 	UTF8* source = *sourceStart;
@@ -349,7 +349,7 @@ ConversionResult ConvertUTF8toUTF16 (
 /* --------------------------------------------------------------------- */
 
 ConversionResult ConvertUTF32toUTF8 (
-		UTF32** sourceStart, const UTF32* sourceEnd, 
+		UTF32** sourceStart, const UTF32* sourceEnd,
 		UTF8** targetStart, const UTF8* targetEnd, const ConversionFlags flags) {
 	ConversionResult result = conversionOK;
 	UTF32* source = *sourceStart;
@@ -358,7 +358,7 @@ ConversionResult ConvertUTF32toUTF8 (
 		UTF32 ch;
 		unsigned short bytesToWrite = 0;
 		const UTF32 byteMask = 0xBF;
-		const UTF32 byteMark = 0x80; 
+		const UTF32 byteMark = 0x80;
 		ch = *source++;
 		/* surrogates of any stripe are not legal UTF32 characters */
 		if (flags == strictConversion ) {
@@ -376,7 +376,7 @@ ConversionResult ConvertUTF32toUTF8 (
 		} else {				bytesToWrite = 2;
 							ch = UNI_REPLACEMENT_CHAR;
 		}
-		
+
 		target += bytesToWrite;
 		if (target > targetEnd) {
 			target -= bytesToWrite; result = targetExhausted; break;
@@ -397,7 +397,7 @@ ConversionResult ConvertUTF32toUTF8 (
 /* --------------------------------------------------------------------- */
 
 ConversionResult ConvertUTF8toUTF32 (
-		UTF8** sourceStart, UTF8* sourceEnd, 
+		UTF8** sourceStart, UTF8* sourceEnd,
 		UTF32** targetStart, const UTF32* targetEnd, const ConversionFlags flags) {
 	ConversionResult result = conversionOK;
 	UTF8* source = *sourceStart;
