+++ developer/src/kmanalyze/kmanalyze.vcxproj
@@ -179,10 +179,10 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\..\common\windows\cpp\include\vkeys.h" />
-    <ClInclude Include="..\..\..\common\windows\delphi\ext\sentry\sentry.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\vkeys.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\delphi\ext\sentry\sentry.h" />
     <ClInclude Include="..\..\..\windows\src\global\inc\keyman64.h" />
-    <ClInclude Include="..\..\..\windows\src\global\inc\keymansentry.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\keymansentry.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
@@ -194,7 +194,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="$(KEYMAN_ROOT)\common\windows\cpp\src\xstring.cpp" />
-    <ClCompile Include="..\..\..\common\windows\cpp\src\vkeys.cpp" />
+    <ClCompile Include="$(KEYMAN_ROOT)\common\windows\cpp\src\vkeys.cpp" />
     <ClCompile Include="kmanalyze.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
