+++ developer/src/kmanalyze/pch.h
@@ -1,4 +1,4 @@
-// Tips for Getting Started: 
+// Tips for Getting Started:
 //   1. Use the Solution Explorer window to add/manage files
 //   2. Use the Team Explorer window to connect to source control
 //   3. Use the Output window to see build output and other messages
@@ -16,6 +16,6 @@
 #include <windows.h>
 #include <string.h>
 #include <stdio.h>
-#include <keymansentry.h>
+#include "../../../common/windows/cpp/include/keymansentry.h"
 #include "keyman64.h"
 #endif //PCH_H
