+++ developer/src/kmcmpdll/kmcmpdll.vcxproj
@@ -288,8 +288,7 @@
   <ItemGroup>
     <ClInclude Include="$(KEYMAN_ROOT)\developer\src\ext\json-schema-validator\json-schema.hpp" />
     <ClInclude Include="$(KEYMAN_ROOT)\developer\src\ext\json\json.hpp" />
-    <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\certificate_check.h" />
-    <ClInclude Include="..\..\..\common\windows\cpp\include\vkeys.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\vkeys.h" />
     <ClInclude Include="comperr.h" />
     <ClInclude Include="compfile.h" />
     <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\legacy_kmx_file.h" />
@@ -299,8 +298,8 @@
     <ClInclude Include="CheckFilenameConsistency.h" />
     <ClInclude Include="CheckNCapsConsistency.h" />
     <ClInclude Include="comprc.h" />
-    <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\ConvertUTF.h" />
-    <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\crc32.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\ConvertUTF.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\crc32.h" />
     <ClInclude Include="debugstore.h" />
     <ClInclude Include="DeprecationChecks.h" />
     <ClInclude Include="Edition.h" />
@@ -309,7 +308,7 @@
     <ClInclude Include="NamedCodeConstants.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="resource.h" />
-    <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\unicode.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\unicode.h" />
     <ClInclude Include="UnreachableRules.h" />
     <ClInclude Include="versioning.h" />
     <ClInclude Include="virtualcharkeys.h" />
@@ -334,7 +333,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="$(KEYMAN_ROOT)\common\windows\cpp\src\xstring.cpp" />
-    <ClCompile Include="..\..\..\common\windows\cpp\src\vkeys.cpp" />
+    <ClCompile Include="$(KEYMAN_ROOT)\common\windows\cpp\src\vkeys.cpp" />
     <ClCompile Include="CasedKeys.cpp" />
     <ClCompile Include="CharToKeyConversion.cpp" />
     <ClCompile Include="CheckForDuplicates.cpp" />
