+++ windows/src/engine/keyman32/keymanengine.h
@@ -266,7 +266,7 @@ void keybd_shift(LPINPUT pInputs, int* n, BOOL isReset, LPBYTE const kbd);
 #include "..\..\..\..\common\windows\cpp\include\keynames.h"
 #include "..\..\..\include\kmtip_guids.h"
 
-#include "crc32.h"
+#include "..\..\..\..\common\windows\cpp\include\crc32.h"
 
 #include "k32_tsf.h"
