+++ windows/src/engine/keyman64/keyman64.vcxproj.filters
@@ -46,12 +46,8 @@
     <ClInclude Include="..\keyman32\appint\appint.h" />
     <ClInclude Include="..\keyman32\calldll.h" />
     <ClInclude Include="..\keyman32\capsstate.h" />
-    <ClInclude Include="..\..\global\inc\certificate_check.h" />
     <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\legacy_kmx_file.h" />
-    <ClInclude Include="..\..\global\inc\ConvertUTF.h" />
-    <ClInclude Include="..\..\global\inc\crc32.h" />
-    <ClInclude Include="..\..\global\inc\crypt_guid.h" />
-    <ClInclude Include="..\..\global\inc\crypt_user.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\crc32.h" />
     <ClInclude Include="..\keyman32\globals.h" />
     <ClInclude Include="..\keyman32\hotkeys.h" />
     <ClInclude Include="..\keyman32\k32_res.h" />
@@ -62,12 +58,8 @@
     <ClInclude Include="..\keyman32\keyman-debug-etw.h" />
     <ClInclude Include="..\keyman32\keyeventsenderthread.h" />
     <ClInclude Include="..\keyman32\keystate.h" />
-    <ClInclude Include="..\..\global\inc\Kmnreg.h" />
     <ClInclude Include="..\keyman32\kmprot.h" />
-    <ClInclude Include="..\..\global\inc\kmprot_hwkey.h" />
-    <ClInclude Include="..\..\global\inc\kmprot_licencenumber.h" />
-    <ClInclude Include="..\..\global\inc\registration.h" />
-    <ClInclude Include="..\..\global\inc\registry.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\registry.h" />
     <ClInclude Include="..\keyman32\pch.h" />
     <ClInclude Include="..\keyman32\security.h" />
     <ClInclude Include="..\keyman32\serialkeyeventclient.h" />
@@ -76,7 +68,7 @@
     <ClInclude Include="..\keyman32\SharedBuffers.h" />
     <ClInclude Include="..\keyman32\syskbd.h" />
     <ClInclude Include="..\keyman32\testkeymanfunctioning.h" />
-    <ClInclude Include="..\..\global\inc\unicode.h" />
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\unicode.h" />
     <ClInclude Include="..\keyman32\vkscancodes.h" />
     <ClInclude Include="..\keyman32\keymanengine.h" />
     <ClInclude Include="..\keyman32\kmprocessactions.h" />
@@ -90,7 +82,7 @@
     <ResourceCompile Include="Keyman64.RC" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="..\..\..\..\common\resources\bitmaps\arrow.bmp" />
-    <None Include="..\..\..\..\common\resources\bitmaps\mask.bmp" />
+    <None Include="$(KEYMAN_ROOT)\common\resources\bitmaps\arrow.bmp" />
+    <None Include="$(KEYMAN_ROOT)\common\resources\bitmaps\mask.bmp" />
   </ItemGroup>
 </Project>
