+++ windows/src/engine/kmtip/kmtip.vcxproj.filters
@@ -90,9 +90,6 @@
     </ResourceCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\global\inc\certificate_check.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\legacy_kmx_file.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -108,7 +105,7 @@
     <ClInclude Include="kmtip.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\global\inc\registry.h">
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\registry.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="pch.h">
@@ -117,7 +114,7 @@
     <ClInclude Include="..\..\..\include\kmtip_guids.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\common\windows\cpp\include\vkeys.h">
+    <ClInclude Include="$(KEYMAN_ROOT)\common\windows\cpp\include\vkeys.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
