+++ windows/src/engine/mcompile/pch.h
@@ -11,7 +11,7 @@
 #include "mcompile.h"
 #include "mc_kmxfile.h"
 #include "mc_syskbd.h"
-#include "../../global/inc/crc32.h"
+#include "../../../../common/windows/cpp/include/crc32.h"
 #include "../../../../common/windows/cpp/include/legacy_kmx_file.h"
-#include "../../global/inc/xstring.h"
-#include "../../global/inc/keymansentry.h"
+#include "../../../../common/windows/cpp/include/xstring.h"
+#include "../../../../common/windows/cpp/include/keymansentry.h"
